wip snapshot, decouple logic, split stdio cli, http server and prepare library entrypoint
