+++ src/examples/clientSse.ts
@@ -23,10 +23,11 @@ const dirname = path.dirname(filename);
 
 dotenv.config({ path: path.resolve(dirname, '../../.env') });
 
-const SERVER_URL = 'https://actors-mcp-server.apify.actor/sse';
+const SERVER_URL = process.env.MCP_SERVER_URL_BASE || 'https://actors-mcp-server.apify.actor/sse';
 // We need to change forward slash / to underscore -- in the tool name as Anthropic does not allow forward slashes in the tool name
 const SELECTED_TOOL = actorNameToToolName('apify/rag-web-browser');
-const QUERY = 'web browser for Anthropic';
+//const QUERY = 'web browser for Anthropic';
+const QUERY = 'apify';
 
 if (!process.env.APIFY_TOKEN) {
     console.error('APIFY_TOKEN is required but not set in the environment variables.');
@@ -98,6 +99,8 @@ async function main(): Promise<void> {
         } else {
             console.error('An unknown error occurred:', error);
         }
+    } finally {
+        await client.close();
     }
 }
